requirements.txt

streamlit>=1.51.0
pandas>=2.3.0
numpy>=1.24.0,<2.1
pillow>=12.0.0
matplotlib>=3.10.0
plotly>=6.0.0
opencv-python>=4.12.0.88

torch>=2.0.0
torchvision>=0.15.0
monai>=1.5.0

langchain-core>=1.0.4
langchain-community>=0.4.1
langchain-openai>=1.0.2
chromadb>=1.3.4

mysql-connector-python>=9.5.0

requests>=2.32.0
python-dotenv>=1.2.0
tqdm>=4.67.0
